Testing integration of average kinematics fortran script

+++ scripts/Prod/createPhysicsList.py
@@ -3,7 +3,7 @@
 #
 # Description:
 # ================================================================
-# Time-stamp: "2023-02-20 05:33:19 trottar"
+# Time-stamp: "2023-02-20 05:36:52 trottar"
 # ================================================================
 #
 # Author:  Richard L. Trotta III <trotta@cua.edu>
@@ -171,7 +171,7 @@ with open(f_list_settings, 'r') as f:
 ################################################################################################################################################
 
 if float(runNumRight[0]) != 0:
-    f_list = '{}/src/kindata/kindata.{}_{}_{}_{:.0f}_-{:.0f}.dat'.format(LTANAPATH, PID, POL, Q2.replace(".",""), float(EPSVAL)*100, thpq_right*1000)
+    f_list = '{}/src/kindata/kindata.{}_{}_{}_{:.0f}_-{:.0f}.dat'.format(LTANAPATH, PID, POL, Q2.replace(".",""), float(EPSVAL)*100, int(thpq_right*1000))
 
     if not os.path.exists(f_list):
         open(f_list, "w").close()    
@@ -187,7 +187,7 @@ if float(runNumRight[0]) != 0:
                     write_to_file(f_list,check_line)
 
 if float(runNumLeft[0]) != 0:
-    f_list = '{}/src/kindata/kindata.{}_{}_{}_{:.0f}_+{:.0f}.dat'.format(LTANAPATH, PID, POL, Q2.replace(".",""), float(EPSVAL)*100, thpq_left*1000)
+    f_list = '{}/src/kindata/kindata.{}_{}_{}_{:.0f}_+{:.0f}.dat'.format(LTANAPATH, PID, POL, Q2.replace(".",""), float(EPSVAL)*100, int(thpq_left*1000))
 
     if not os.path.exists(f_list):
         open(f_list, "w").close()
